:new: stellordnung
Inspired by a functional programming style that every function should be one expression, I tried to rephrse a whole paragraph into a long sentence,
which reduces readablity compared to nested bullets but gives me some confidence that I do not need to split this slip to several slips.

Thanks to GPT-4.1 for proofreading.