:new: page data variable `removal`
Add a new variable `removal` in page data.
If its value is truthy, then a "marked for removal" message will be displayed in the page header,
titled `RM: {reason}` where reason is the value of `removal` variable.

All pages marked for removal will be deleted in the future commit after the next tagged commit,
the url will redirect to the corresponding source file at that tag.